bkup: semi-refactor.
