+++ white_rc_car_manipulator/core/inc/usart.h
@@ -50,7 +50,7 @@
 #ifndef __usart_H
 #define __usart_H
 #ifdef __cplusplus
- extern "C" {
+extern "C" {
 #endif
 
 /* Includes ------------------------------------------------------------------*/
