+++ white_rc_car_manipulator/src/lib_system.c
@@ -27,48 +27,54 @@
 
 #include "startup.h"
 
+//#ifdef USER_TASK1
+//extern osThreadId task1_t;
+//#endif
+//#ifdef USER_TASK2
+//extern osThreadId task2_t;
+//#endif
+//#ifdef USER_TASK3
+//extern osThreadId task3_t;
+//#endif
+//#ifdef USER_TASK4
+//extern osThreadId task4_t;
+//#endif
+//#ifdef USER_TASK5
+//extern osThreadId task5_t;
+//#endif
 
-#ifdef USER_TASK1
-  extern osThreadId task1_t;
-#endif
-#ifdef USER_TASK2
-  extern osThreadId task2_t;
-#endif
-#ifdef USER_TASK3
-  extern osThreadId task3_t;
-#endif
-#ifdef USER_TASK4
-  extern osThreadId task4_t;
-#endif
-#ifdef USER_TASK5
-  extern osThreadId task5_t;
-#endif
+extern osThreadId task1_t, task2_t;
 
 
 void sys_start_task(void) {
-#ifdef USER_TASK1
     osThreadDef(ostask1, USER_TASK1, osPriorityAboveNormal, 0, 128);
     task1_t = osThreadCreate(osThread(ostask1), NULL);
-#endif
 
-#ifdef USER_TASK2
     osThreadDef(ostask2, USER_TASK2, osPriorityAboveNormal, 0, 128);
     task2_t = osThreadCreate(osThread(ostask2), NULL);
-#endif
 
-#ifdef USER_TASK3
-    osThreadDef(ostask3, USER_TASK3, osPriorityNormal, 0, 128);
-    task3_t = osThreadCreate(osThread(ostask3), NULL);
-#endif
-
-#ifdef USER_TASK4
-    osThreadDef(ostask4, USER_TASK4, osPriorityNormal, 0, 128);
-    task4_t = osThreadCreate(osThread(ostask4), NULL);
-#endif
-
-#ifdef USER_TASK5
-    osThreadDef(ostask5, USER_TASK5, osPriorityNormal, 0, 128);
-    task5_t = osThreadCreate(osThread(ostask5), NULL);
-#endif
-
+//#ifdef USER_TASK1
+//    osThreadDef(ostask1, USER_TASK1, osPriorityAboveNormal, 0, 128);
+//    task1_t = osThreadCreate(osThread(ostask1), NULL);
+//#endif
+//
+//#ifdef USER_TASK2
+//    osThreadDef(ostask2, USER_TASK2, osPriorityAboveNormal, 0, 128);
+//    task2_t = osThreadCreate(osThread(ostask2), NULL);
+//#endif
+//
+//#ifdef USER_TASK3
+//    osThreadDef(ostask3, USER_TASK3, osPriorityNormal, 0, 128);
+//    task3_t = osThreadCreate(osThread(ostask3), NULL);
+//#endif
+//
+//#ifdef USER_TASK4
+//    osThreadDef(ostask4, USER_TASK4, osPriorityNormal, 0, 128);
+//    task4_t = osThreadCreate(osThread(ostask4), NULL);
+//#endif
+//
+//#ifdef USER_TASK5
+//    osThreadDef(ostask5, USER_TASK5, osPriorityNormal, 0, 128);
+//    task5_t = osThreadCreate(osThread(ostask5), NULL);
+//#endif
 }
